2023-12-06 part 2
Note that this is the rewritten version which uses math.
The naive solution ran in like 250ms which is *fine* I guess, but *once* in a while I knew how to do this.
I mean, on a basic level.
I forgot my entire school math years ago.
I had to look tons of stuff up but I did manage to look things up and piece it together myself without having to use a solver.
So here we go, this is me doing math for the first time in forever (probably the last time too).

Signed-off-by: benaryorg <binary@benary.org>